refactor(cognition): extract the gym grader into one shared module
test_grade / grade_rust + helpers were private to cognition/eval.rs. The
teacher-episode generator (genome/teach, next slice) must grade write→fix
trajectories with the EXACT same verdict the A/B evaluator uses, or a gene
that "passes" generation could "fail" eval for grader-shape reasons alone.

Per the compression principle (one logical decision, one place): lift the
grader into cognition/gym_grader.rs with pub test_grade + extract_code_block,
its tests move with it, eval.rs now `use`s it. No behavior change.
